Fix warnings hidden earlier due to including application headers as system includes

+++ vespamalloc/src/vespamalloc/util/callstack.h
@@ -75,7 +75,7 @@ size_t StackEntry<StackRep>::fillStack(StackEntry<StackRep> *stack, size_t nelem
     } else {
         sz = 0;
     }
-    if (sz < nelems) {
+    if (size_t(sz) < nelems) {
         stack[sz] = StackEntry<StackRep>();
     }
     return sz;
